Move TransactionError from the cli to the core
