+++ cli/src/lib.rs
@@ -19,7 +19,6 @@
 
 use lazy_static::lazy_static;
 use radicle_registry_client::*;
-use std::convert::TryInto;
 use structopt::StructOpt;
 use thiserror::Error as ThisError;
 
@@ -146,23 +145,4 @@ pub enum CommandError {
 
     #[error(transparent)]
     AccountStorageError(#[from] account_storage::Error),
-}
-
-/// The subset of possible errors having led a transaction to failure.
-#[derive(Debug, ThisError)]
-pub enum TransactionError {
-    #[error(transparent)]
-    RegistryError(RegistryError),
-
-    #[error("{0:?}")]
-    OtherDispatchError(DispatchError),
-}
-
-impl From<DispatchError> for TransactionError {
-    fn from(dispatch_error: DispatchError) -> Self {
-        dispatch_error
-            .try_into()
-            .map(TransactionError::RegistryError)
-            .unwrap_or(TransactionError::OtherDispatchError(dispatch_error))
-    }
 }
